requirements.txt
asgiref==3.8.1
distlib==0.3.9
dj-database-url==3.0.1
Django==5.2.3
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
filelock==3.17.0
graphviz==0.21
gunicorn==23.0.0
packaging==25.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
PyJWT==2.9.0
sqlparse==0.5.3
tzdata==2025.2
virtualenv==20.29.2
whitenoise==6.9.0
